requirements.txt
langchain
fastapi
uvicorn
python-dotenv
twilio
openai
python-multipart
deeplake
pypdf
doctran